+++ Resources/JCudaVectorValueEdition.cu
@@ -1,5 +1,5 @@
 extern "C"
-__global__ void changeVectorValue(int position, float value,float *vector)
+__global__ void changeVectorValue(int position, double value,double *vector)
 {
     vector[position]=value;
 };
